Move Work Carousel to Home folder
