+++ src/components/Home/TrendingWorks/TrendingWorks.tsx
@@ -3,7 +3,7 @@ import {Alert, View} from 'react-native';
 import WorksApi from '../../../api/openlibrary/works';
 import Work from '../../../api/openlibrary/works/model/Work';
 import Text from '../../Text';
-import WorkCarousel from '../../WorkCarousel';
+import WorkCarousel from '../WorkCarousel';
 import styles from './TrendingWorks.styles';
 import {TrendingWorksProps} from './TrendingWorks.types';
