+++ src/components/Home/WorkCarousel/WorkCarousel.styles.ts
