+++ src/components/Home/WorkCarousel/WorkCarousel.types.ts
@@ -1,0 +1,3 @@
+import Work from '../../../api/openlibrary/works/model/Work';
+
+export type WorkCarouselProps = {works: Work[]};
