--- src/components/WorkCarousel/WorkCarousel.types.ts
@@ -1,3 +1,0 @@
-import Work from '../../api/openlibrary/works/model/Work';
-
-export type WorkCarouselProps = {works: Work[]};
