+++ src/screens/Home/HomeScreen.tsx
@@ -4,7 +4,7 @@ import {SafeAreaView} from 'react-native-safe-area-context';
 import WorksApi from '../../api/openlibrary/works';
 import Work from '../../api/openlibrary/works/model/Work';
 import TrendingWorks from '../../components/Home/TrendingWorks';
-import WorkCarousel from '../../components/WorkCarousel';
+import WorkCarousel from '../../components/Home/WorkCarousel';
 
 import styles from './HomeScreen.styles';
 import {HomeScreenProps} from './HomeScreen.types';
